Update to SDL3

+++ main/game.cpp
@@ -779,7 +779,7 @@ MAC(extern uint8_t joydefs_calibrating;)
 
 void game_flush_inputs()
 {
-	int dx,dy;
+	float dx,dy;
 	key_flush();
 	joy_flush();
 	mouse_flush();
